+++ src/Catesta/Resources/Vault/src/Tests/v5/Unit/ExportedFunctions.Tests.ps1
@@ -1,17 +1,14 @@
-#-------------------------------------------------------------------------
-Set-Location -Path $PSScriptRoot
-#-------------------------------------------------------------------------
-$ModuleName = '<%=$PLASTER_PARAM_ModuleName%>'
-$PathToManifest = [System.IO.Path]::Combine('..', '..', $ModuleName, "$ModuleName.psd1")
-#-------------------------------------------------------------------------
-if (Get-Module -Name $ModuleName -ErrorAction 'SilentlyContinue') {
-    #if the module is already in memory, remove it
-    Remove-Module -Name $ModuleName -Force
-}
-Import-Module $PathToManifest -Force
-#-------------------------------------------------------------------------
-
 BeforeAll {
+    Set-Location -Path $PSScriptRoot
+    $ModuleName = '<%=$PLASTER_PARAM_ModuleName%>'
+    $PathToManifest = [System.IO.Path]::Combine('..', '..', $ModuleName, "$ModuleName.psd1")
+    Get-Module $ModuleName -ErrorAction SilentlyContinue | Remove-Module -Force
+    Import-Module $PathToManifest -Force
+    $manifestContent = Test-ModuleManifest -Path $PathToManifest
+    $moduleExported = Get-Command -Module $ModuleName | Select-Object -ExpandProperty Name
+    $manifestExported = ($manifestContent.ExportedFunctions).Keys
+}
+BeforeDiscovery {
     Set-Location -Path $PSScriptRoot
     $ModuleName = '<%=$PLASTER_PARAM_ModuleName%>'
     $PathToManifest = [System.IO.Path]::Combine('..', '..', $ModuleName, "$ModuleName.psd1")
@@ -24,21 +21,20 @@ Describe $ModuleName {
     Context 'Exported Commands' -Fixture {
 
         Context 'Number of commands' -Fixture {
-            It -Name 'Exports the same number of public functions as what is listed in the Module Manifest' -Test {
+
+            It 'Exports the same number of public functions as what is listed in the Module Manifest' {
                 $manifestExported.Count | Should -BeExactly $moduleExported.Count
             }
+
         }
 
-        Context 'Explicitly exported commands' -ForEach $moduleExported {
-            foreach ($command in $moduleExported) {
-                BeforeAll {
-                    $command = $_
-                }
-                It -Name "Includes the $command in the Module Manifest ExportedFunctions" -Test {
-                    $manifestExported -contains $command | Should -BeTrue
-                }
+        Context 'Explicitly exported commands' {
+
+            It 'Includes <_> in the Module Manifest ExportedFunctions' -ForEach $moduleExported {
+                $manifestExported -contains $_ | Should -BeTrue
             }
+
         }
-    }
+    } #context_ExportedCommands
 
 }