+++ src/Tests/Integration/FileChecks.Tests.ps1
@@ -1,10 +1,10 @@
-#-------------------------------------------------------------------------
-Set-Location -Path $PSScriptRoot
-#-------------------------------------------------------------------------
-$ModuleName = 'Catesta'
-$resourcePath1 = [System.IO.Path]::Combine( '..', '..', 'Artifacts', 'Resources')
-$manifests = Get-ChildItem -Path $resourcePath1 -Include '*.xml' -Recurse
-#-------------------------------------------------------------------------
+BeforeAll {
+    Set-Location -Path $PSScriptRoot
+    $ModuleName = 'Catesta'
+    $resourcePath1 = [System.IO.Path]::Combine( '..', '..', 'Artifacts', 'Resources')
+    $manifests = Get-ChildItem -Path $resourcePath1 -Include '*.xml' -Recurse
+}
+
 Describe 'File Checks' {
     BeforeAll {
         $WarningPreference = 'Continue'
@@ -73,6 +73,10 @@ Describe 'File Checks' {
 
         It 'should have a private function example' {
             $srcFiles.Name.Contains('Get-Day.ps1') | Should -BeExactly $true
+        } #it
+
+        It 'should have a class example' {
+            $srcFiles.Name.Contains('SampleClass.ps1') | Should -BeExactly $true
         } #it
 
     } #context_module